Move ModalInteraction into correct package
